Draft: Resolve "Feature Request: PostgreSQL"

Closes #562

See merge request pub/terrareg!444

+++ terrareg/alembic/versions/a36ffbb6580e_add_index_to_parent_module_version_.py
@@ -8,6 +8,8 @@ Create Date: 2022-06-25 13:03:41.700763
 from alembic import op
 import sqlalchemy as sa
 
+from terrareg.alembic.versions import constraint_exists
+
 
 # revision identifiers, used by Alembic.
 revision = 'a36ffbb6580e'
@@ -17,28 +19,13 @@ depends_on = None
 
 
 def upgrade():
-    try:
-        # Attempt to remove the index and re-create excitily, as this would be left over from
-        # the old foreign key
-        with op.batch_alter_table('analytics', schema=None) as batch_op:
+    with op.batch_alter_table('analytics', schema=None) as batch_op:
+        if constraint_exists(op.get_bind(), 'analytics', 'fk_analytics_parent_module_version_module_version_id'):
             batch_op.drop_index('fk_analytics_parent_module_version_module_version_id')
-            batch_op.create_index(op.f('ix_analytics_parent_module_version'), ['parent_module_version'], unique=False)
-    except (sa.exc.OperationalError, sa.exc.ProgrammingError):
-        # For SQLite, the index would no longer exist, so simply create without deleting
-        # the original
-        with op.batch_alter_table('analytics', schema=None) as batch_op:
-            batch_op.create_index(op.f('ix_analytics_parent_module_version'), ['parent_module_version'], unique=False)
-
+        batch_op.create_index(op.f('ix_analytics_parent_module_version'), ['parent_module_version'], unique=False)
 
 def downgrade():
-    try:
-        # Attempt to remove the index and re-create excitily, as this would be left over from
-        # the old foreign key
-        with op.batch_alter_table('analytics', schema=None) as batch_op:
+    with op.batch_alter_table('analytics', schema=None) as batch_op:
+        if constraint_exists(op.get_bind(), 'analytics', 'ix_analytics_parent_module_version'):
             batch_op.drop_index(op.f('ix_analytics_parent_module_version'))
-            batch_op.create_index('fk_analytics_parent_module_version_module_version_id', ['parent_module_version'], unique=False)
-    except sa.exc.OperationalError:
-        # For SQLite, the index would no longer exist, so simply create without deleting
-        # the original
-        with op.batch_alter_table('analytics', schema=None) as batch_op:
-            batch_op.create_index('fk_analytics_parent_module_version_module_version_id', ['parent_module_version'], unique=False)
+        batch_op.create_index('fk_analytics_parent_module_version_module_version_id', ['parent_module_version'], unique=False)
